+++ src/builtins/mod.rs
@@ -1810,9 +1810,13 @@ pub fn setup_full(vm: &mut Vm) -> error::Result<()> {
         "DataView",
         data_view_constructor,
         &[
+            ("getInt16", data_view_get_int16, 1),
             ("getInt8", data_view_get_int8, 1),
+            ("getUint16", data_view_get_uint16, 1),
             ("getUint8", data_view_get_uint8, 1),
+            ("setInt16", data_view_set_int16, 2),
             ("setInt8", data_view_set_int8, 2),
+            ("setUint16", data_view_set_uint16, 2),
             ("setUint8", data_view_set_uint8, 2),
         ],
     )?;